REFACTOR: Split properties + tests in more files
